requirements.txt
asgiref==3.8.1
Django==5.0.7
django-allauth==0.63.6
django-allauth-2fa==0.11.1
django-otp==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
packaging==24.1
PyJWT==2.8.0
pyotp==2.9.0
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.5.1
typing_extensions==4.12.2
